2.25.1

+++ internal-links/core/links/text-to-link-converter-factory.php
@@ -23,7 +23,7 @@ final class Text_To_Link_Converter_Factory
         $elapsed_time_in_secs = time() - intval($_SERVER['REQUEST_TIME']);
         $available_time_in_secs = intval(ini_get('max_execution_time')) - $elapsed_time_in_secs;
         // use 70 percent of available time.
-        $time_in_secs = round(0.7 * $available_time_in_secs, 2);
+        $time_in_secs = (int) round(0.7 * $available_time_in_secs, 2);
         /**
          * There can be two scenarios
          *
@@ -67,9 +67,9 @@ final class Text_To_Link_Converter_Factory
         if ($cache_option) {
             // Put a caching layer over link builder.
             $with_cache = new Cache_Layer($id, $type, $link_builder);
-            return new Timeout_Monitor_Layer(self::get_time_required_for_content_linking($default_allowed_time_in_secs), $with_cache);
+            return new Timeout_Monitor_Layer(self::get_time_required_for_content_linking((int) $default_allowed_time_in_secs), $with_cache);
         } else {
-            return new Timeout_Monitor_Layer(self::get_time_required_for_content_linking($default_allowed_time_in_secs), $link_builder);
+            return new Timeout_Monitor_Layer(self::get_time_required_for_content_linking((int) $default_allowed_time_in_secs), $link_builder);
         }
     }
 }
